\name{plot.hcd}
\alias{plot.hcd}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Plot a Hollow Curve Distribution. }
\description{
Basic method to plot a hollow curve distribution from an 'hcd'object'.}
\usage{
plot.hcd(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{An object of class 'hcd'.}
  \item{\dots}{Extra arguments to be passed to \code{plot}.}
}
\details{
The precise plot depends on which components are present in the 'hcd' object. The simplest call to \code{hcd.generate} will produce a simple line plot showing the average clade richness for the taxa in rank order of richness. If the argument \code{minmax} was set to TRUE then the minimum and maximum diversity at each rank are shown. If the 'hcd' object was generated by a call to \code{hcd.fit} then the hollow curve distribution of the real data is shown.
}
\author{David Orme}
\note{ Needs expanding to allow users to control the plot through \code{...} arguments.}

\seealso{\code{\link{hcd.generate}}, \code{\link{hcd.fit}}}
\examples{
par(mfrow=c(2,2))

hcdExample <- hcd.generate(50,20)
plot(hcdExample)

hcdExample <- hcd.generate(50,20, minmax=TRUE)
plot(hcdExample)

primates <- c(80, 19, 15, 10, 10, 9, 9, 7, 7, 6, 6, 5, 5, 5, 4, 3, 1, 1, 1)
names(primates) <- c("Cercopithecidae","Callitrichidae","Aotus & Callicebus","Pithecia & Cacajao","Galigidae","Hylobates","Lemurinae","Atelini","Cheirogalidae","Aloutta","Indroidea","Pongidae & Homo","Saimiri","Lorisidae","Tarsius","other Cebus","Cebus apella","Lepilemur","Varecia")

hcdPrimate <- hcd.fit(primates)
plot(hcdPrimate)

hcdPrimate <- hcd.fit(primates, minmax=TRUE)
plot(hcdPrimate)
}
\keyword{ hplot }% at least one, from doc/KEYWORDS
\keyword{ distribution }% __ONLY ONE__ keyword per line
